HOME-321 - Login page layouting
