+++ smarthome/constants/colors.swift
@@ -1,0 +1,14 @@
+//
+//  constants.swift
+//  smarthome
+//
+//  Created by Oskar on 15/01/2021.
+//
+
+import SwiftUI
+import UIKit
+
+struct CColors {
+    static let blue = Color(UIColor.systemBlue)
+    static let gray = Color(UIColor(red: 200/255, green: 200/255, blue: 200/255, alpha: 1))
+}
