+++ smarthome/constants/sizings.swift
@@ -1,0 +1,11 @@
+//
+//  sizings.swift
+//  smarthome
+//
+//  Created by Oskar on 15/01/2021.
+//
+import SwiftUI
+
+struct CSizings {
+    static let rounding: CGFloat = 6.0
+}
